Improve aggregation data for Rating Statistics

+++ src/main/java/com/mafia/statistics/MafiaStatisticsAPI/dto/statistics/additional/TopRatingTable.java
@@ -17,5 +17,5 @@ public class TopRatingTable {
 
     private String playerNickname;
 
-    private Float points;
+    private Double points;
 }
